+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `"utilityGridView"` option to the `Icon` component.
+
 ## 4.1.0 (November 20, 2025)
 
 ### Adds